+++ pyproject.toml
@@ -549,6 +549,7 @@ markers = [
     "athena: mark a test as AWS Athena-dependent.",
     "spark: mark a test as Spark-dependent.",
     "postgres: mark a test as postgres-dependent.",
+    "mysql: mark a test as mysql-dependent.",
     "project: mark a test that verifies properties of the gx project",
     "pyarrow: mark a test as PyArrow-dependent.",
     "openpyxl: mark a test for openpyxl-dependent, which is for Excel files.",
