+++ tests/datasource/data_connector/test_configured_asset_gcs_data_connector.py
@@ -25,6 +25,9 @@ if not google.storage:
         allow_module_level=True,
     )
 
+# module level markers
+pytestmark = [pytest.mark.external_cloud_client]
+
 
 @pytest.fixture
 def expected_config_dict():
