+++ tests/datasource/data_connector/test_data_connector.py
@@ -10,7 +10,6 @@ from great_expectations.core.batch import (
 from great_expectations.core.yaml_handler import YAMLHandler
 from great_expectations.data_context.util import instantiate_class_from_config
 from great_expectations.datasource.data_connector import (
-    ConfiguredAssetFilesystemDataConnector,
     InferredAssetFilesystemDataConnector,
 )
 from great_expectations.datasource.data_connector.util import (
@@ -53,77 +52,6 @@ assets:
     return basic_data_connector
 
 
-# TODO: <Alex>This test should be moved to "tests/datasource/data_connector/test_configured_asset_filesystem_data_connector.py".</Alex>
-def test_basic_instantiation(tmp_path_factory):
-    base_directory = str(
-        tmp_path_factory.mktemp("basic_data_connector__filesystem_data_connector")
-    )
-
-    # noinspection PyUnusedLocal
-    my_data_connector = ConfiguredAssetFilesystemDataConnector(  # noqa: F841
-        name="my_data_connector",
-        datasource_name="FAKE_DATASOURCE",
-        execution_engine=PandasExecutionEngine(),
-        base_directory=base_directory,
-        glob_directive="*.csv",
-        default_regex={
-            "pattern": "(.*)",
-            "group_names": ["file_name"],
-        },
-        assets={"my_asset_name": {}},
-    )
-
-
-# TODO: <Alex>This test should be potentially moved to "tests/datasource/data_connector/test_configured_asset_filesystem_data_connector.py".</Alex>
-def test__get_instantiation_through_instantiate_class_from_config(basic_data_connector):
-    # noinspection PyProtectedMember
-    data_references: list = (
-        basic_data_connector._get_data_reference_list_from_cache_by_data_asset_name(
-            data_asset_name="my_asset_name"
-        )
-    )
-    assert len(data_references) == 0
-    assert data_references == []
-
-
-# TODO: <Alex>This test should be renamed properly and moved to "tests/datasource/data_connector/test_configured_asset_filesystem_data_connector.py".</Alex>
-def test__file_object_caching_for_FileDataConnector(tmp_path_factory):
-    base_directory = str(
-        tmp_path_factory.mktemp("basic_data_connector__filesystem_data_connector")
-    )
-    create_files_in_directory(
-        directory=base_directory,
-        file_name_list=[
-            "pretend/path/A-100.csv",
-            "pretend/path/A-101.csv",
-            "pretend/directory/B-1.csv",
-            "pretend/directory/B-2.csv",
-        ],
-    )
-
-    my_data_connector = ConfiguredAssetFilesystemDataConnector(
-        name="my_data_connector",
-        datasource_name="FAKE_DATASOURCE",
-        execution_engine=PandasExecutionEngine(),
-        base_directory=base_directory,
-        glob_directive="*/*/*.csv",
-        default_regex={
-            "pattern": "(.*).csv",
-            "group_names": ["name"],
-        },
-        assets={"stuff": {}},
-    )
-
-    assert my_data_connector.get_data_reference_count() == 0
-    assert len(my_data_connector.get_unmatched_data_references()) == 0
-
-    # noinspection PyProtectedMember
-    my_data_connector._refresh_data_references_cache()
-
-    assert len(my_data_connector.get_unmatched_data_references()) == 0
-    assert my_data_connector.get_data_reference_count() == 4
-
-
 def test_get_batch_definition_list_from_batch_request():
     pass
 
@@ -157,6 +85,7 @@ def test_available_data_asset_names():
 
 
 # TODO: <Alex>This test should be moved to the test module that is dedicated to BatchRequest and BatchDefinition testing.</Alex>
+@pytest.mark.unit
 def test__batch_definition_matches_batch_request():
     # TODO: <Alex>We need to cleanup PyCharm warnings.</Alex>
     A = BatchDefinition(
@@ -250,6 +179,7 @@ def test__batch_definition_matches_batch_request():
     # TODO : Test cases to exercise ranges, etc.
 
 
+@pytest.mark.filesystem
 def test_for_self_check_using_InferredAssetFilesystemDataConnector_PandasExecutionEngine(
     tmp_path_factory,
 ):
@@ -281,6 +211,7 @@ def test_for_self_check_using_InferredAssetFilesystemDataConnector_PandasExecuti
     # assert self_check_results["example_data_reference"]["n_rows"] == 2
 
 
+@pytest.mark.spark
 def test_for_self_check_using_InferredAssetFilesystemDataConnector_SparkDFExecutionEngine(
     spark_session, basic_spark_df_execution_engine, tmp_path_factory
 ):
