+++ tests/datasource/data_connector/test_inferred_asset_filesystem_data_connector.py
@@ -17,6 +17,9 @@ from tests.test_utils import create_files_in_directory
 
 yaml = YAMLHandler()
 
+# module level markers
+pytestmark = [pytest.mark.filesystem]
+
 
 def test_basic_instantiation(tmp_path_factory):
     base_directory = str(tmp_path_factory.mktemp("test_basic_instantiation"))
