+++ tests/datasource/data_connector/test_inferred_asset_gcs_data_connector.py
@@ -20,6 +20,9 @@ if not google.storage:
         allow_module_level=True,
     )
 
+# module level markers
+pytestmark = [pytest.mark.external_cloud_client]
+
 
 @pytest.fixture
 def expected_config_dict():
