+++ tests/datasource/data_connector/test_inferred_asset_glue_catalog_data_connector.py
@@ -14,6 +14,9 @@ from great_expectations.validator.validator import Validator
 
 yaml = YAMLHandler()
 
+# module level markers
+pytestmark = [pytest.mark.external_cloud_client]
+
 
 def test_basic_instantiation(glue_titanic_catalog):
     random.seed(0)
