+++ tests/datasource/data_connector/test_sql_data_connector.py
@@ -28,6 +28,7 @@ yaml = YAMLHandler()
 
 
 # TODO: <Alex>Some methods in this module are misplaced and/or provide no action; this must be repaired.</Alex>
+@pytest.mark.sqlite
 def test_basic_self_check(test_cases_for_sql_data_connector_sqlite_execution_engine):
     random.seed(0)
     execution_engine = test_cases_for_sql_data_connector_sqlite_execution_engine
@@ -82,6 +83,7 @@ def test_basic_self_check(test_cases_for_sql_data_connector_sqlite_execution_eng
     }
 
 
+@pytest.mark.sqlite
 def get_data_context_for_datasource_and_execution_engine(
     context: AbstractDataContext,
     connection_url: str,
@@ -110,6 +112,7 @@ def get_data_context_for_datasource_and_execution_engine(
     return context
 
 
+@pytest.mark.sqlite
 @pytest.mark.parametrize("splitter_method_name_prefix", ["_", ""])
 def test_get_batch_definition_list_from_batch_request(
     splitter_method_name_prefix,
@@ -212,6 +215,7 @@ def test_get_batch_definition_list_from_batch_request(
         )
 
 
+@pytest.mark.sqlite
 def test_example_A(test_cases_for_sql_data_connector_sqlite_execution_engine):
     random.seed(0)
     execution_engine = test_cases_for_sql_data_connector_sqlite_execution_engine
@@ -266,6 +270,7 @@ def test_example_A(test_cases_for_sql_data_connector_sqlite_execution_engine):
     }
 
 
+@pytest.mark.sqlite
 @pytest.mark.parametrize("splitter_method_name_prefix", ["_", ""])
 def test_example_B(
     splitter_method_name_prefix,
@@ -323,6 +328,7 @@ def test_example_B(
     }
 
 
+@pytest.mark.sqlite
 @pytest.mark.parametrize("splitter_method_name_prefix", ["_", ""])
 def test_example_C(
     splitter_method_name_prefix,
@@ -383,6 +389,7 @@ def test_example_C(
     }
 
 
+@pytest.mark.sqlite
 def test_example_E(test_cases_for_sql_data_connector_sqlite_execution_engine):
     random.seed(0)
     execution_engine = test_cases_for_sql_data_connector_sqlite_execution_engine
@@ -436,6 +443,7 @@ def test_example_E(test_cases_for_sql_data_connector_sqlite_execution_engine):
     }
 
 
+@pytest.mark.sqlite
 def test_example_F(test_cases_for_sql_data_connector_sqlite_execution_engine):
     random.seed(0)
     execution_engine = test_cases_for_sql_data_connector_sqlite_execution_engine
@@ -489,6 +497,7 @@ def test_example_F(test_cases_for_sql_data_connector_sqlite_execution_engine):
     }
 
 
+@pytest.mark.sqlite
 @pytest.mark.parametrize("splitter_method_name_prefix", ["_", ""])
 def test_example_G(
     splitter_method_name_prefix,
@@ -553,6 +562,7 @@ def test_example_G(
     }
 
 
+@pytest.mark.sqlite
 def test_example_H(test_cases_for_sql_data_connector_sqlite_execution_engine):
     return
 
@@ -736,29 +746,7 @@ def test_get_batch_data_and_markers_sampling_method__a_list(
     assert len(validator.head(fetch_all=True)) == 4
 
 
-def test_get_batch_data_and_markers_sampling_method__md5(
-    test_cases_for_sql_data_connector_sqlite_execution_engine,
-):
-    # noinspection PyUnusedLocal
-    execution_engine = (  # noqa: F841
-        test_cases_for_sql_data_connector_sqlite_execution_engine
-    )
-
-    # SQlite doesn't support MD5
-    # batch_data, batch_markers = execution_engine.get_batch_data_and_markers(
-    #     batch_spec=SqlAlchemyDatasourceBatchSpec({
-    #         "table_name": "table_partitioned_by_date_column__A",
-    #         "batch_identifiers": {},
-    #         "splitter_method": "_split_on_whole_table",
-    #         "splitter_kwargs": {},
-    #         "sampling_method": "_sample_using_md5",
-    #         "sampling_kwargs": {
-    #             "column_name": "index",
-    #         }
-    #     })
-    # )
-
-
+@pytest.mark.sqlite
 def test_get_batch_data_and_markers_to_make_sure_splitter_and_sampler_methods_are_optional(
     in_memory_runtime_context,
     test_cases_for_sql_data_connector_sqlite_connection_url,
@@ -1206,6 +1194,7 @@ def test_default_behavior_with_no_splitter(
     assert batch_definition_list[0]["batch_identifiers"] == {}
 
 
+@pytest.mark.sqlite
 def test_behavior_with_whole_table_splitter(
     test_cases_for_sql_data_connector_sqlite_execution_engine,
 ):
@@ -1267,6 +1256,7 @@ def test_behavior_with_whole_table_splitter(
     assert batch_definition_list[0]["batch_identifiers"] == {}
 
 
+@pytest.mark.sqlite
 def test_basic_instantiation_of_InferredAssetSqlDataConnector(
     test_cases_for_sql_data_connector_sqlite_execution_engine,
 ):
@@ -1356,6 +1346,7 @@ def test_basic_instantiation_of_InferredAssetSqlDataConnector(
     assert len(batch_definition_list) == 1
 
 
+@pytest.mark.sqlite
 def test_more_complex_instantiation_of_InferredAssetSqlDataConnector(
     test_cases_for_sql_data_connector_sqlite_execution_engine,
 ):
@@ -1444,6 +1435,7 @@ def test_more_complex_instantiation_of_InferredAssetSqlDataConnector(
     assert len(batch_definition_list) == 1
 
 
+@pytest.mark.sqlite
 def test_basic_instantiation_of_ConfiguredAssetSqlDataConnector(
     test_cases_for_sql_data_connector_sqlite_execution_engine,
 ):
@@ -1475,6 +1467,7 @@ def test_basic_instantiation_of_ConfiguredAssetSqlDataConnector(
     }
 
 
+@pytest.mark.sqlite
 @pytest.mark.parametrize("splitter_method_name_prefix", ["_", ""])
 def test_more_complex_instantiation_of_ConfiguredAssetSqlDataConnector(
     splitter_method_name_prefix,
@@ -1517,6 +1510,7 @@ def test_more_complex_instantiation_of_ConfiguredAssetSqlDataConnector(
     }
 
 
+@pytest.mark.sqlite
 @mock.patch("great_expectations.execution_engine.SqlAlchemyExecutionEngine.__init__")
 @pytest.mark.parametrize("splitter_method_name_prefix", ["_", ""])
 def test_more_complex_instantiation_of_ConfiguredAssetSqlDataConnector_include_schema_name(
@@ -1555,6 +1549,7 @@ def test_more_complex_instantiation_of_ConfiguredAssetSqlDataConnector_include_s
     assert "table_partitioned_by_date_column__A" in my_data_connector.assets
 
 
+@pytest.mark.sqlite
 @mock.patch("great_expectations.execution_engine.SqlAlchemyExecutionEngine.__init__")
 @pytest.mark.parametrize("splitter_method_name_prefix", ["_", ""])
 def test_more_complex_instantiation_of_ConfiguredAssetSqlDataConnector_include_schema_name_prefix_suffix(
@@ -1583,6 +1578,7 @@ def test_more_complex_instantiation_of_ConfiguredAssetSqlDataConnector_include_s
     )
 
 
+@pytest.mark.sqlite
 @mock.patch("great_expectations.execution_engine.SqlAlchemyExecutionEngine.__init__")
 @pytest.mark.parametrize("splitter_method_name_prefix", ["_", ""])
 def test_more_complex_instantiation_of_ConfiguredAssetSqlDataConnector_include_schema_name_prefix_suffix_table_name(
@@ -1609,6 +1605,7 @@ def test_more_complex_instantiation_of_ConfiguredAssetSqlDataConnector_include_s
     assert "taxi__main.my_asset__asset" in my_data_connector.assets
 
 
+@pytest.mark.sqlite
 @pytest.mark.parametrize("splitter_method_name_prefix", ["_", ""])
 def test_more_complex_instantiation_of_ConfiguredAssetSqlDataConnector_include_schema_name_prefix_suffix_table_name_asset_splitter(
     splitter_method_name_prefix,
@@ -1647,6 +1644,7 @@ def test_more_complex_instantiation_of_ConfiguredAssetSqlDataConnector_include_s
     assert len(batch_definition_list) == 34
 
 
+@pytest.mark.sqlite
 @pytest.mark.parametrize("splitter_method_name_prefix", ["_", ""])
 def test_more_complex_instantiation_of_ConfiguredAssetSqlDataConnector_include_schema_name_prefix_suffix_table_name_default_splitter(
     splitter_method_name_prefix,
@@ -1687,6 +1685,7 @@ def test_more_complex_instantiation_of_ConfiguredAssetSqlDataConnector_include_s
     assert len(batch_definition_list) == 34
 
 
+@pytest.mark.sqlite
 @pytest.mark.parametrize("splitter_method_name_prefix", ["_", ""])
 def test_more_complex_instantiation_of_ConfiguredAssetSqlDataConnector_include_schema_name_prefix_suffix_table_name_default_splitter_asset_override(
     splitter_method_name_prefix,
